requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
requests==2.3.0
static3==0.5.1
whitenoise==1.0.4
boto==2.34.0
Pillow==2.6.1
